CMake: Simplify install on Windows
The install BINDIR, SBINDIR, LIBDIR, and INCLUDEDIR may be set in
the top level CMakeLists.txt and does not need to be set for each
target. This also enables a build to customize those so that the
bins COULD go to the "bin" directory.

Also add CMake option "ENABLE_WINDOWS_INSTALL_THIRDPARTY_DEPENDENCIES".
- When enabled (default), the install will include required system runtime
  libraries and also a copy of the other third party library dependencies
  (e.g. libpcre2, libxml2, libcrypto, etc.).
- When disabled, those files are not included.
  One situation where this is useful is if vcpkg is being used to build
  ClamAV itself (including its dependencies).