Replace CL_CLEAN with CL_SUCCESS in clamav.h, and scanners.c
Both enum variants are 0, so it's a no-op.

`cl_error_t` should be used to determine if we keep going or stop, whether
that's because there was a detection and we're not in allmatch mode, or if
because of an error.

`cl_verdict_t` should be used to determine or say the verdict (clean is a
verdict, though I feel 'nothing found' is more accurate).