libclamav: Fix issue scanning with no signatures loaded (#1560)
If the application never calls `cl_load()`, then the clean-cache is
never initialized. That is a legitimate mode to run in when perhaps we
just want to extract stuff, record metadata, or for fuzzing.

This commit adds a check if the ctx->engine->cache is NULL. If it is,
then we treat it as though caching is disabled.

CLAM-2856