ZIP: Fix possible leak (#1568)
Fix a possible memory leak in the overlapping files detecting logic.
The issue is because cleanup for the zip catalogue allocated by this
function only happens if the status is no CL_SUCCESS.

My fix uses a better pattern to ensure we don't override a format error
with a "clean" result when adding the heuristic alert.

Fixes: https://issues.oss-fuzz.com/issues/376723678

CLAM-2857