Windows: Fix issue creating new temp subdirectory
On Windows, when scanning some files with `--leave-temps` enabled and
with `--tempdir` set to something like `C:\temp`, it may fail to create
a new subdirectory to store the temp files because the absolute file
path is too long for the `mkdir()` function.

The `mkdir()` function may fail on Windows if the filepath is longer than
the legacy MAX_PATH.
Fixing this in C or C++ is rather difficult, requiring either a registry
key + application manifest change, or else converting the path to UTF16
and UNC format (i.e. `"\\?\C:\temp"`) to pass to `CreateDirectoryW()`.

The solution in this commit is to use the Rust `std` library instead. It
is able to handle the longer file paths without issue.

CLAM-2924