libclamav: Malformed regexes result in misleading memory error messages.
The regex parsing code treated anything that did not return CL_SUCCESS
as CL_EMEM in several places, resulting in a return code indicative of a
malformed database returning a memory error instead. This change passes
the real return code up the chain, and also prevents trying to realloc 0
bytes when we don't have anything to realloc due to a malformed
database.

CLAM-2191