Use non deprecated OpenSSL accessors for ASN1_STRING values.
This changes your time conversion routine to use
ASN1_STRING_get0_data() and ASN1_STRING_length() rather than
directly using the values from the object.

OpenSSL plans to make ASN1_STRING opaque soon:
https://github.com/openssl/openssl/issues/29117

The accessor in question is in OpenSSL 1.1 as well as
BoringSSL and LibreSSL so should be widely available.

As the returned value is const, and you are changing it
to pass to strptime, I've made it make a copy to work with.

Since I added another allocation I converted it to single
return to ensure everything gets freed appropriately.